requirements.txt
apache-airflow-providers-google[all]
apache-airflow-providers-amazon[all]